LayoutTests/imported/w3c/web-platform-tests/selection/shadow-dom/tentative/Selection-getComposedRanges-dom-mutations-removal_mode=closed-expected.txt

slotted content 2

FAIL Range is fully in shadow tree. Removing shadow host collapses composed StaticRange. Note it does not update previously returned composed StaticRange. assert_equals: expected 1 but got 2
PASS Range is fully in shadow tree. Removing parent of shadow host collapses composed StaticRange.
PASS Range is in light DOM. Removing startContainer rescopes new composed range to its parent.
PASS Range is across shadow trees. Replacing shadowRoot content rescopes new composed range to the shadowRoot.
PASS Range is across shadow trees. Removing ancestor shadow host rescopes composed range end to parent.
PASS Range is between two light slotted contents. Removing start container rescopes to its parent in light tree.
